pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pulsewriter"
version = "0.1.0"
description = "PulseWriter — write once, publish everywhere (Core + CLI + API)."
readme = "README.md"
requires-python = ">=3.10"
authors = [{name="Nikita Koselev"}]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.7.0",
    "jinja2>=3.1.4",
    "pyyaml>=6.0.2",
    "typer>=0.12.3",
    "httpx>=0.27.0"
]

[project.optional-dependencies]
dev = ["pytest>=8.2.0", "pytest-cov>=5.0.0"]

[project.scripts]
pulsewriter = "pulsewriter_cli.main:app"

[tool.setuptools]
packages = ["pulsewriter_core", "pulsewriter_cli", "pulsewriter_api"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
pulsewriter_core = ["templates/*.j2"]
